+++ packages/storage/src/index.ts
@@ -5,4 +5,3 @@ export * from './mail-rules.js';
 export * from './transport-rules.js';
 export * from './itip-inbound.js';
 export * from './protocol-cert.js';
-export * from './oauth-clients-bootstrap.js';